+++ app/urls.py
@@ -22,10 +22,12 @@ from content_post.apis.v1.comment_router import content as comment_router
 from content_post.apis.v1.detail_router import content as detail_router
 from content_post.apis.v1.main_router import content as main_router
 from content_post.apis.v1.test_create_feed_router import content as create_router
+from content_post.apis.v1.scrap_router import content as scrap_router
 from user_admission import views
 from user_admission.apis.v1.login_router import account as login_router
 from user_admission.apis.v1.logout_router import account as logout_router
 from user_admission.apis.v1.register_router import account as register_router
+
 
 # version 1.0.0 >> 중요한 변경 / 중간 변경 / 최소 변경
 
@@ -35,6 +37,7 @@ api.add_router("login/", login_router)
 api.add_router("register/", register_router)
 api.add_router("", main_router)
 api.add_router("detail/", detail_router)
+api.add_router("detail/", scrap_router)
 api.add_router("", create_router)
 api.add_router("logout/", logout_router)
 api.add_router("comment/", comment_router)
